requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.6
idna==3.2
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2021.1
redis==3.5.3
requests==2.26.0
six==1.16.0
tornado==6.1
tzlocal==3.0
urllib3==1.26.7
vk-api==11.9.4
